refactor: extract mods toolbar into its own widget; drop dead card
Splits the 4145-line mods_screen.dart down to ~3550:
- Removes the unused private _ModCardWidget (the old card design,
  superseded by the ModCardWidget component) — ~265 lines of dead code.
- Extracts the search/sort/tag-filter/favorites toolbar into a
  self-contained ModsToolbar ConsumerWidget (components/mods_toolbar.dart),
  now that its state lives in providers. It owns the search controller and
  tag-menu overlay controllers. A new clearModFilters(ref) helper in
  state_providers is shared by the toolbar and the no-results screen.
- Removes a stale unused keybinds_widget import.

No behaviour change.
