+++ mod_manager_flutter/lib/screens/mods_screen.dart
@@ -25,7 +25,7 @@ import '../l10n/app_localizations.dart';
 import 'components/mode_toggle_widget.dart';
 import 'components/character_cards_list_widget.dart';
 import 'components/mod_card_widget.dart';
-import 'components/keybinds_widget.dart';
+import 'components/mods_toolbar.dart';
 
 /// A staged gallery entry in the edit dialog. It's one of: an image already in
 /// the mod folder ([existingPath]), a newly picked file to import on save
@@ -87,13 +87,7 @@ class _ModsScreenState extends ConsumerState<ModsScreen>
   late final ScrollController _modsScrollController = ScrollController();
 
   // Mods list sorting & filtering live in providers (see state_providers.dart);
-  // only the search box keeps a local controller, mirrored to its provider.
-  final TextEditingController _searchController = TextEditingController();
-
-  // Anchored dropdown for the tag filter.
-  final OverlayPortalController _tagMenuController = OverlayPortalController();
-  final LayerLink _tagMenuLink = LayerLink();
-
+  // the toolbar UI lives in [ModsToolbar].
   bool get _isFiltering => ref.read(modFiltersActiveProvider);
 
   // Collapsed group ids in the grouped "ALL" view (in-memory, by character id).
@@ -102,9 +96,6 @@ class _ModsScreenState extends ConsumerState<ModsScreen>
   @override
   void initState() {
     super.initState();
-    // The search query persists in a provider; restore the box's text if this
-    // screen is rebuilt while a query is active.
-    _searchController.text = ref.read(modSearchQueryProvider);
     _loadingAnimationController = AnimationController(
       duration: const Duration(milliseconds: 1500),
       vsync: this,
@@ -138,10 +129,8 @@ class _ModsScreenState extends ConsumerState<ModsScreen>
     _characterSelectionDebounce?.cancel();
     _focusNode.dispose();
     _modsScrollController.dispose();
-    _searchController.dispose();
     super.dispose();
   }
-
 
   Future<void> _loadTags() async {
     final configService = await ApiService.getConfigService();
@@ -2269,11 +2258,10 @@ class _ModsScreenState extends ConsumerState<ModsScreen>
               ],
             ),
           ),
-          // Search / sort / tag-filter toolbar. Scoped in its own Consumer so a
-          // keystroke or filter toggle rebuilds only the toolbar, not the
-          // whole screen.
-          if (currentSkins.isNotEmpty)
-            Consumer(builder: (context, ref, _) => _buildModsToolbar(ref)),
+          // Search / sort / tag-filter toolbar. Self-contained and
+          // provider-driven, so a keystroke or filter toggle rebuilds only the
+          // toolbar, not the whole screen.
+          if (currentSkins.isNotEmpty) const ModsToolbar(),
 
           // Counter for active mods
           if (currentSkins.isNotEmpty)
@@ -2732,320 +2720,6 @@ class _ModsScreenState extends ConsumerState<ModsScreen>
     );
   }
 
-  void _clearFilters() {
-    _searchController.clear();
-    ref.read(modSearchQueryProvider.notifier).state = '';
-    ref.read(modTagFiltersProvider.notifier).state = <String>{};
-    ref.read(modFavoritesOnlyProvider.notifier).state = false;
-  }
-
-  String _sortLabel(ModSort mode) {
-    switch (mode) {
-      case ModSort.added:
-        return loc.t('mods.sort.added');
-      case ModSort.nameAsc:
-        return loc.t('mods.sort.name_asc');
-      case ModSort.nameDesc:
-        return loc.t('mods.sort.name_desc');
-    }
-  }
-
-  /// Search field + sort menu + tag filter menu + favorites toggle, all on one
-  /// compact row above the mods grid.
-  Widget _buildModsToolbar(WidgetRef ref) {
-    final tags = ref.watch(availableModTagsProvider);
-    final searchQuery = ref.watch(modSearchQueryProvider);
-    final isFiltering = ref.watch(modFiltersActiveProvider);
-    return Padding(
-      padding: EdgeInsets.fromLTRB(
-        AppConstants.defaultPadding,
-        AppConstants.smallPadding,
-        AppConstants.defaultPadding,
-        0,
-      ),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Row(
-            children: [
-              Expanded(
-                child: TextField(
-                  controller: _searchController,
-                  onChanged: (v) =>
-                      ref.read(modSearchQueryProvider.notifier).state = v,
-                  decoration: InputDecoration(
-                    isDense: true,
-                    hintText: loc.t('mods.toolbar.search'),
-                    prefixIcon: const Icon(Icons.search, size: 18),
-                    suffixIcon: searchQuery.isNotEmpty
-                        ? IconButton(
-                            icon: const Icon(Icons.close, size: 16),
-                            tooltip: loc.t('mods.toolbar.clear'),
-                            onPressed: () {
-                              _searchController.clear();
-                              ref.read(modSearchQueryProvider.notifier).state =
-                                  '';
-                            },
-                          )
-                        : null,
-                    contentPadding: const EdgeInsets.symmetric(
-                      horizontal: 12,
-                      vertical: 10,
-                    ),
-                    border: OutlineInputBorder(
-                      borderRadius: BorderRadius.circular(10),
-                    ),
-                  ),
-                ),
-              ),
-              const SizedBox(width: 8),
-              _buildSortButton(ref),
-              if (tags.isNotEmpty) ...[
-                const SizedBox(width: 8),
-                _buildTagFilterButton(ref, tags),
-              ],
-              const SizedBox(width: 8),
-              _buildFavoritesToggle(ref),
-            ],
-          ),
-          if (isFiltering)
-            Align(
-              alignment: Alignment.centerLeft,
-              child: TextButton.icon(
-                onPressed: _clearFilters,
-                icon: const Icon(Icons.clear, size: 16),
-                label: Text(loc.t('mods.toolbar.clear_filters')),
-                style: TextButton.styleFrom(
-                  padding: const EdgeInsets.symmetric(horizontal: 8),
-                  visualDensity: VisualDensity.compact,
-                ),
-              ),
-            ),
-        ],
-      ),
-    );
-  }
-
-  Widget _toolbarButton({required Widget child, bool active = false}) {
-    return Container(
-      padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 11),
-      decoration: BoxDecoration(
-        color: active ? const Color(0xFF0EA5E9).withOpacity(0.12) : null,
-        borderRadius: BorderRadius.circular(10),
-        border: Border.all(
-          color: active
-              ? const Color(0xFF0EA5E9).withOpacity(0.6)
-              : Colors.grey.withOpacity(0.4),
-        ),
-      ),
-      child: child,
-    );
-  }
-
-  Widget _buildSortButton(WidgetRef ref) {
-    final sortMode = ref.watch(modSortProvider);
-    return PopupMenuButton<ModSort>(
-      tooltip: loc.t('mods.toolbar.sort'),
-      initialValue: sortMode,
-      onSelected: (m) => ref.read(modSortProvider.notifier).state = m,
-      itemBuilder: (_) => [
-        for (final m in ModSort.values)
-          CheckedPopupMenuItem(
-            value: m,
-            checked: sortMode == m,
-            child: Text(_sortLabel(m)),
-          ),
-      ],
-      child: _toolbarButton(
-        child: Row(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            const Icon(Icons.sort, size: 18),
-            const SizedBox(width: 6),
-            Text(_sortLabel(sortMode), style: const TextStyle(fontSize: 13)),
-            const Icon(Icons.arrow_drop_down, size: 18),
-          ],
-        ),
-      ),
-    );
-  }
-
-  /// Tag filter as an anchored dropdown panel (scales to many tags) with an
-  /// Any/All match-mode toggle. Built with OverlayPortal so it's a normal part
-  /// of the tree — toggles update state directly and re-filter live.
-  Widget _buildTagFilterButton(WidgetRef ref, List<String> tags) {
-    // Count only tags present in this view — matches what visibleModsProvider
-    // actually applies, so the badge can't claim a filter that does nothing.
-    final count = ref.watch(modTagFiltersProvider).where(tags.contains).length;
-    return OverlayPortal(
-      controller: _tagMenuController,
-      overlayChildBuilder: (context) {
-        return Stack(
-          children: [
-            // Tap-outside barrier to dismiss.
-            Positioned.fill(
-              child: GestureDetector(
-                behavior: HitTestBehavior.opaque,
-                onTap: _tagMenuController.hide,
-              ),
-            ),
-            // Anchor the panel's right edge to the button's right edge so it
-            // grows leftward into the window instead of clipping off the right.
-            CompositedTransformFollower(
-              link: _tagMenuLink,
-              targetAnchor: Alignment.bottomRight,
-              followerAnchor: Alignment.topRight,
-              offset: const Offset(0, 6),
-              child: _buildTagMenuPanel(tags),
-            ),
-          ],
-        );
-      },
-      child: CompositedTransformTarget(
-        link: _tagMenuLink,
-        child: InkWell(
-          borderRadius: BorderRadius.circular(10),
-          onTap: _tagMenuController.toggle,
-          child: _toolbarButton(
-            active: count > 0,
-            child: Row(
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                const Icon(Icons.filter_list, size: 18),
-                const SizedBox(width: 6),
-                Text(loc.t('mods.toolbar.tags'), style: const TextStyle(fontSize: 13)),
-                if (count > 0) ...[
-                  const SizedBox(width: 6),
-                  Container(
-                    padding: const EdgeInsets.symmetric(horizontal: 6, vertical: 1),
-                    decoration: BoxDecoration(
-                      color: const Color(0xFF0EA5E9),
-                      borderRadius: BorderRadius.circular(8),
-                    ),
-                    child: Text(
-                      '$count',
-                      style: const TextStyle(fontSize: 11, color: Colors.white, fontWeight: FontWeight.bold),
-                    ),
-                  ),
-                ],
-                const Icon(Icons.arrow_drop_down, size: 18),
-              ],
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-
-  Widget _buildTagMenuPanel(List<String> tags) {
-    // The panel lives in the app Overlay (via OverlayPortal), so it watches the
-    // filter providers through its own Consumer to rebuild on toggle.
-    return Consumer(
-      builder: (context, ref, _) {
-        final activeTags = ref.watch(modTagFiltersProvider);
-        final matchAll = ref.watch(modTagMatchAllProvider);
-        return Material(
-          elevation: 8,
-          borderRadius: BorderRadius.circular(10),
-          clipBehavior: Clip.antiAlias,
-          child: SizedBox(
-            width: 280,
-            child: Column(
-              mainAxisSize: MainAxisSize.min,
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Padding(
-                  padding: const EdgeInsets.fromLTRB(14, 12, 14, 8),
-                  child: Row(
-                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                    children: [
-                      Text(
-                        loc.t('mods.toolbar.match'),
-                        style: const TextStyle(
-                          fontSize: 12,
-                          fontWeight: FontWeight.w600,
-                        ),
-                      ),
-                      Wrap(
-                        spacing: 6,
-                        children: [
-                          ChoiceChip(
-                            label: Text(loc.t('mods.toolbar.match_any')),
-                            selected: !matchAll,
-                            visualDensity: VisualDensity.compact,
-                            onSelected: (_) => ref
-                                .read(modTagMatchAllProvider.notifier)
-                                .state = false,
-                          ),
-                          ChoiceChip(
-                            label: Text(loc.t('mods.toolbar.match_all')),
-                            selected: matchAll,
-                            visualDensity: VisualDensity.compact,
-                            onSelected: (_) => ref
-                                .read(modTagMatchAllProvider.notifier)
-                                .state = true,
-                          ),
-                        ],
-                      ),
-                    ],
-                  ),
-                ),
-                const Divider(height: 1),
-                ConstrainedBox(
-                  constraints: const BoxConstraints(maxHeight: 280),
-                  child: ListView(
-                    shrinkWrap: true,
-                    padding: EdgeInsets.zero,
-                    children: [
-                      for (final tag in tags)
-                        CheckboxListTile(
-                          dense: true,
-                          controlAffinity: ListTileControlAffinity.leading,
-                          visualDensity: VisualDensity.compact,
-                          value: activeTags.contains(tag),
-                          title: Text(tag, style: const TextStyle(fontSize: 13)),
-                          onChanged: (sel) {
-                            final next = Set<String>.from(activeTags);
-                            if (sel == true) {
-                              next.add(tag);
-                            } else {
-                              next.remove(tag);
-                            }
-                            ref.read(modTagFiltersProvider.notifier).state =
-                                next;
-                          },
-                        ),
-                    ],
-                  ),
-                ),
-              ],
-            ),
-          ),
-        );
-      },
-    );
-  }
-
-  Widget _buildFavoritesToggle(WidgetRef ref) {
-    final favoritesOnly = ref.watch(modFavoritesOnlyProvider);
-    return Tooltip(
-      message: loc.t('mods.toolbar.favorites_only'),
-      child: InkWell(
-        borderRadius: BorderRadius.circular(10),
-        onTap: () =>
-            ref.read(modFavoritesOnlyProvider.notifier).state = !favoritesOnly,
-        child: _toolbarButton(
-          active: favoritesOnly,
-          child: Icon(
-            favoritesOnly ? Icons.star : Icons.star_border,
-            size: 18,
-            color: favoritesOnly ? const Color(0xFFFACC15) : null,
-          ),
-        ),
-      ),
-    );
-  }
-
   Widget _buildNoResults() {
     return Center(
       child: Column(
@@ -3059,7 +2733,7 @@ class _ModsScreenState extends ConsumerState<ModsScreen>
           ),
           const SizedBox(height: 12),
           TextButton.icon(
-            onPressed: _clearFilters,
+            onPressed: () => clearModFilters(ref),
             icon: const Icon(Icons.clear, size: 18),
             label: Text(loc.t('mods.toolbar.clear_filters')),
           ),
@@ -3874,272 +3548,5 @@ class _ModsScreenState extends ConsumerState<ModsScreen>
         );
       }
     }
-  }
-}
-
-class _ModCardWidget extends StatefulWidget {
-  final ModInfo mod;
-  final bool isDarkMode;
-  final Map<String, String> modCharacterTags;
-  final String Function(String) getCharacterName;
-  final LinearGradient Function(ModInfo, bool, bool) getModCardGradient;
-  final Color Function(ModInfo, bool, bool) getModCardBorderColor;
-  final List<BoxShadow> Function(ModInfo, bool, bool) getModCardShadows;
-
-  const _ModCardWidget({
-    required this.mod,
-    required this.isDarkMode,
-    required this.modCharacterTags,
-    required this.getCharacterName,
-    required this.getModCardGradient,
-    required this.getModCardBorderColor,
-    required this.getModCardShadows,
-  });
-
-  @override
-  State<_ModCardWidget> createState() => _ModCardWidgetState();
-}
-
-class _ModCardWidgetState extends State<_ModCardWidget> {
-  bool isHovered = false;
-
-  @override
-  Widget build(BuildContext context) {
-    return MouseRegion(
-      onEnter: (_) => setState(() => isHovered = true),
-      onExit: (_) => setState(() => isHovered = false),
-      child: AnimatedContainer(
-        duration: const Duration(milliseconds: 300),
-        curve: Curves.easeOutCubic,
-        transform: Matrix4.identity()
-          ..scale(isHovered ? 1.02 : 1.0)
-          ..translate(0.0, isHovered ? -4.0 : 0.0),
-        decoration: BoxDecoration(
-          borderRadius: BorderRadius.circular(20),
-          gradient: widget.getModCardGradient(
-            widget.mod,
-            widget.isDarkMode,
-            isHovered,
-          ),
-          border: Border.all(
-            color: widget.getModCardBorderColor(
-              widget.mod,
-              widget.isDarkMode,
-              isHovered,
-            ),
-            width: widget.mod.isActive ? 2.5 : (isHovered ? 2.0 : 1.2),
-          ),
-          boxShadow: widget.getModCardShadows(
-            widget.mod,
-            widget.isDarkMode,
-            isHovered,
-          ),
-        ),
-        child: Container(
-          decoration: BoxDecoration(
-            borderRadius: BorderRadius.circular(19),
-            gradient: isHovered
-                ? LinearGradient(
-                    begin: Alignment.topLeft,
-                    end: Alignment.bottomRight,
-                    colors: [
-                      Color.fromRGBO(255, 255, 255, 0.05),
-                      Colors.transparent,
-                    ],
-                  )
-                : null,
-          ),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.stretch,
-            children: [
-              // Зображення моду з новим стилем
-              Expanded(
-                child: ClipRRect(
-                  borderRadius: const BorderRadius.vertical(
-                    top: Radius.circular(18),
-                  ),
-                  child: Container(
-                    decoration: BoxDecoration(
-                      gradient:
-                          widget.mod.imagePath != null &&
-                              File(widget.mod.imagePath!).existsSync()
-                          ? null
-                          : LinearGradient(
-                              begin: Alignment.topLeft,
-                              end: Alignment.bottomRight,
-                              colors: [
-                                widget.isDarkMode
-                                    ? const Color(0xFF374151)
-                                    : const Color(0xFFF3F4F6),
-                                widget.isDarkMode
-                                    ? const Color(0xFF1F2937)
-                                    : const Color(0xFFE5E7EB),
-                              ],
-                            ),
-                    ),
-                    child: Stack(
-                      fit: StackFit.expand,
-                      children: [
-                        if (widget.mod.imagePath != null &&
-                            File(widget.mod.imagePath!).existsSync())
-                          AnimatedContainer(
-                            duration: const Duration(milliseconds: 300),
-                            child: Image.file(
-                              File(widget.mod.imagePath!),
-                              fit: BoxFit.cover,
-                              key: ValueKey(
-                                '${widget.mod.id}_${widget.mod.imagePath}',
-                              ),
-                              cacheWidth: null,
-                              cacheHeight: null,
-                            ),
-                          )
-                        else
-                          Center(
-                            child: Container(
-                              padding: const EdgeInsets.all(20),
-                              decoration: BoxDecoration(
-                                color: widget.isDarkMode
-                                    ? Color.fromRGBO(255, 255, 255, 0.05)
-                                    : Color.fromRGBO(0, 0, 0, 0.03),
-                                borderRadius: BorderRadius.circular(12),
-                              ),
-                              child: Icon(
-                                Icons.image_outlined,
-                                size: 40,
-                                color: widget.isDarkMode
-                                    ? Color.fromRGBO(255, 255, 255, 0.4)
-                                    : Color.fromRGBO(0, 0, 0, 0.4),
-                              ),
-                            ),
-                          ),
-
-                        // Градієнт оверлей для кращої читабельності
-                        if (widget.mod.imagePath != null &&
-                            File(widget.mod.imagePath!).existsSync())
-                          Container(
-                            decoration: BoxDecoration(
-                              gradient: LinearGradient(
-                                begin: Alignment.topCenter,
-                                end: Alignment.bottomCenter,
-                                colors: [
-                                  Colors.transparent,
-                                  Color.fromRGBO(0, 0, 0, 0.1),
-                                ],
-                              ),
-                            ),
-                          ),
-
-                        // Стильний статус індикатор
-                        Positioned(
-                          top: 12,
-                          right: 12,
-                          child: AnimatedContainer(
-                            duration: const Duration(milliseconds: 200),
-                            padding: const EdgeInsets.all(8),
-                            decoration: BoxDecoration(
-                              color: widget.mod.isActive
-                                  ? const Color(0xFF10B981)
-                                  : widget.isDarkMode
-                                  ? const Color(0xFF374151)
-                                  : const Color(0xFF6B7280),
-                              borderRadius: BorderRadius.circular(12),
-                              boxShadow: [
-                                BoxShadow(
-                                  color: widget.mod.isActive
-                                      ? Color.fromRGBO(16, 185, 129, 0.3)
-                                      : Color.fromRGBO(0, 0, 0, 0.3),
-                                  blurRadius: 8,
-                                  offset: const Offset(0, 2),
-                                ),
-                              ],
-                            ),
-                            child: Icon(
-                              widget.mod.isActive
-                                  ? Icons.check_rounded
-                                  : Icons.close_rounded,
-                              size: 18,
-                              color: Colors.white,
-                            ),
-                          ),
-                        ),
-
-                        // Тег персонажа з новим стилем
-                        if (widget.modCharacterTags.containsKey(widget.mod.id))
-                          Positioned(
-                            top: 12,
-                            left: 12,
-                            child: AnimatedContainer(
-                              duration: const Duration(milliseconds: 200),
-                              padding: const EdgeInsets.symmetric(
-                                horizontal: 10,
-                                vertical: 4,
-                              ),
-                              decoration: BoxDecoration(
-                                color: const Color(0xFF0EA5E9),
-                                borderRadius: BorderRadius.circular(8),
-                                boxShadow: [
-                                  BoxShadow(
-                                    color: Color.fromRGBO(14, 165, 233, 0.4),
-                                    blurRadius: 8,
-                                    offset: const Offset(0, 2),
-                                  ),
-                                ],
-                              ),
-                              child: Text(
-                                widget.getCharacterName(
-                                  widget.modCharacterTags[widget.mod.id]!,
-                                ),
-                                style: const TextStyle(
-                                  color: Colors.white,
-                                  fontSize: 11,
-                                  fontWeight: FontWeight.w700,
-                                  letterSpacing: 0.3,
-                                ),
-                              ),
-                            ),
-                          ),
-                      ],
-                    ),
-                  ),
-                ),
-              ),
-
-              // Назва моду з новим стилем
-              Container(
-                padding: const EdgeInsets.all(16),
-                decoration: BoxDecoration(
-                  borderRadius: const BorderRadius.vertical(
-                    bottom: Radius.circular(18),
-                  ),
-                  color: widget.mod.isActive
-                      ? (widget.isDarkMode
-                            ? Color.fromRGBO(14, 165, 233, 0.1)
-                            : Color.fromRGBO(14, 165, 233, 0.05))
-                      : (widget.isDarkMode
-                            ? Color.fromRGBO(255, 255, 255, 0.02)
-                            : Color.fromRGBO(0, 0, 0, 0.01)),
-                ),
-                child: Text(
-                  widget.mod.name,
-                  style: TextStyle(
-                    fontSize: 14,
-                    fontWeight: FontWeight.w600,
-                    letterSpacing: 0.2,
-                    color: widget.mod.isActive
-                        ? const Color(0xFF0EA5E9)
-                        : (widget.isDarkMode
-                              ? Color.fromRGBO(255, 255, 255, 0.9)
-                              : Color.fromRGBO(0, 0, 0, 0.8)),
-                  ),
-                  maxLines: 2,
-                  overflow: TextOverflow.ellipsis,
-                ),
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
   }
 }
